# openapi.yaml
openapi: 3.0.3
info:
  title: AI Cost Calculator API
  description: |
    The world's first AI model cost estimation API.
    Calculate, compare, and optimize AI spending across all major providers
    including Anthropic, OpenAI, Google, Meta, and Mistral.
  version: 1.0.0
  contact:
    name: lazymac2x
servers:
  - url: http://localhost:5000
    description: Local development
paths:
  /api/v1/calculate:
    post:
      summary: Calculate cost for a single request
      description: Calculate the exact cost of a single AI API call given model, input tokens, and output tokens.
      tags: [Cost Calculation]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [model, input_tokens, output_tokens]
              properties:
                model:
                  type: string
                  example: claude-sonnet-4
                input_tokens:
                  type: integer
                  example: 1000
                output_tokens:
                  type: integer
                  example: 500
                batch:
                  type: boolean
                  default: false
      responses:
        "200":
          description: Cost calculation result
          content:
            application/json:
              schema:
                type: object
                properties:
                  model:
                    type: string
                  model_name:
                    type: string
                  provider:
                    type: string
                  pricing_type:
                    type: string
                  input_tokens:
                    type: integer
                  output_tokens:
                    type: integer
                  input_cost_usd:
                    type: number
                  output_cost_usd:
                    type: number
                  total_cost_usd:
                    type: number
                  rates:
                    type: object
                    properties:
                      input_per_mtok:
                        type: number
                      output_per_mtok:
                        type: number
        "400":
          description: Invalid request
        "404":
          description: Unknown model

  /api/v1/estimate:
    post:
      summary: Estimate monthly cost
      description: Project monthly AI spending based on daily usage patterns.
      tags: [Cost Calculation]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [model, requests_per_day, avg_input_tokens, avg_output_tokens]
              properties:
                model:
                  type: string
                  example: gpt-4o
                requests_per_day:
                  type: integer
                  example: 1000
                avg_input_tokens:
                  type: integer
                  example: 2000
                avg_output_tokens:
                  type: integer
                  example: 500
                batch:
                  type: boolean
                  default: false
                days:
                  type: integer
                  default: 30
      responses:
        "200":
          description: Monthly cost estimate
        "400":
          description: Invalid request
        "404":
          description: Unknown model

  /api/v1/compare:
    post:
      summary: Compare all models
      description: Compare cost across all available AI models for the same workload, sorted cheapest first.
      tags: [Comparison]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [input_tokens, output_tokens]
              properties:
                input_tokens:
                  type: integer
                  example: 5000
                output_tokens:
                  type: integer
                  example: 1000
                batch:
                  type: boolean
                  default: false
                providers:
                  type: array
                  items:
                    type: string
                  example: [openai, anthropic]
      responses:
        "200":
          description: Model comparison results

  /api/v1/cheapest:
    get:
      summary: Find cheapest model
      description: Find the cheapest AI model for given token counts, optionally filtered by budget.
      tags: [Comparison]
      parameters:
        - name: input_tokens
          in: query
          required: true
          schema:
            type: integer
        - name: output_tokens
          in: query
          required: true
          schema:
            type: integer
        - name: budget
          in: query
          schema:
            type: number
          description: Maximum cost per request in USD
        - name: providers
          in: query
          schema:
            type: string
          description: Comma-separated provider filter
      responses:
        "200":
          description: Cheapest model recommendations

  /api/v1/models:
    get:
      summary: List all models
      description: List all supported AI models with full pricing information.
      tags: [Models]
      responses:
        "200":
          description: All models with pricing

  /api/v1/models/{provider}:
    get:
      summary: List models by provider
      description: List all models from a specific provider.
      tags: [Models]
      parameters:
        - name: provider
          in: path
          required: true
          schema:
            type: string
            enum: [anthropic, openai, google, meta, mistral]
      responses:
        "200":
          description: Provider models
        "404":
          description: Unknown provider

  /api/v1/tokens:
    post:
      summary: Estimate token count
      description: Estimate token count from raw text input using character-based heuristic.
      tags: [Utilities]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [text]
              properties:
                text:
                  type: string
                  example: "Hello, how much would this cost to process with GPT-4?"
      responses:
        "200":
          description: Token estimation result
        "400":
          description: Missing text